# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Consultas Médicas",
        "description": "Documentação da API do Sistema de Consultas Médicas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/consultas": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Lista todas as consultas",
                "description": "Retorna uma lista com todas as consultas cadastradas.",
                "operationId": "4de317507609369d17e797e2691793ee",
                "responses": {
                    "200": {
                        "description": "Consultas encontradas",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "id": 20,
                                            "paciente_id": 20,
                                            "medico_id": 1,
                                            "tipo_consulta_id": 1,
                                            "data_atendimento": "2026-06-20T13:00:00.000000Z",
                                            "valor_consulta": "180.00",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z",
                                            "paciente": {
                                                "id": 20,
                                                "nome": "Isabela Cardoso",
                                                "cpf": "00011122234",
                                                "data_nascimento": "1998-08-15T03:00:00.000000Z",
                                                "telefone": "48900000009",
                                                "created_at": "2026-07-22T23:07:46.000000Z",
                                                "updated_at": "2026-07-22T23:07:46.000000Z"
                                            },
                                            "medico": {
                                                "id": 1,
                                                "nome": "Dra. Ana Martins",
                                                "crm": "CRM001",
                                                "created_at": "2026-07-22T23:07:46.000000Z",
                                                "updated_at": "2026-07-22T23:07:46.000000Z"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Cadastra uma nova consulta",
                "description": "Cria uma nova consulta informando paciente, médico, tipo, data, horário e valor.",
                "operationId": "54c6b97ec9124d69f1e6496147efa174",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paciente_id",
                                    "medico_id",
                                    "tipo_consulta_id",
                                    "data_atendimento",
                                    "hora_atendimento",
                                    "valor_consulta"
                                ],
                                "properties": {
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_consulta_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_atendimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-20"
                                    },
                                    "hora_atendimento": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "13:00"
                                    },
                                    "valor_consulta": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 180
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/consultas/{id}": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Busca uma consulta pelo ID",
                "description": "Retorna uma consulta específica pelo seu identificador.",
                "operationId": "a196278920293d1c3dd71dde45ae865e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": {
                                        "id": 20,
                                        "paciente_id": 20,
                                        "medico_id": 1,
                                        "tipo_consulta_id": 1,
                                        "data_atendimento": "2026-06-20T13:00:00.000000Z",
                                        "valor_consulta": "180.00",
                                        "created_at": "2026-07-22T23:07:46.000000Z",
                                        "updated_at": "2026-07-22T23:07:46.000000Z",
                                        "paciente": {
                                            "id": 20,
                                            "nome": "Isabela Cardoso",
                                            "cpf": "00011122234",
                                            "telefone": "48900000009",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        },
                                        "medico": {
                                            "id": 1,
                                            "nome": "Dra. Ana Martins",
                                            "crm": "CRM001",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        },
                                        "tipo_consulta": {
                                            "id": 1,
                                            "nome": "Consulta Cardiológica",
                                            "duracao": 30,
                                            "hora_inicio": "08:00:00",
                                            "hora_fim": "18:00:00",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:07:45.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Atualiza uma consulta",
                "description": "Atualiza uma consulta existente informando paciente, médico, tipo, data, horário e valor.",
                "operationId": "9de29f3a913d1d92bdae7e0a27b97988",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paciente_id",
                                    "medico_id",
                                    "tipo_consulta_id",
                                    "data_atendimento",
                                    "hora_atendimento",
                                    "valor_consulta"
                                ],
                                "properties": {
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_consulta_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "data_atendimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-25"
                                    },
                                    "hora_atendimento": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "15:30"
                                    },
                                    "valor_consulta": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta atualizada com sucesso"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Remove uma consulta",
                "operationId": "b1232cdae57aee2fca324dcc7d8064e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta removida"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                }
            }
        },
        "/api/consultas/medicos-por-tipo/{id}": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Lista médicos por tipo de consulta",
                "description": "Retorna os médicos disponíveis para um determinado tipo de consulta.",
                "operationId": "ab641c6235a01eb8ac576741a7927aae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de médicos"
                    },
                    "404": {
                        "description": "Tipo de consulta não encontrado"
                    }
                }
            }
        },
...
    },
...
}